requirements.txt
# RQ1 & RQ2 Experiment Dependencies

# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Statistics & ML
scipy>=1.10.0
scikit-learn>=1.2.0

# RQ2: Embedding-based Clustering (optional, for B4)
sentence-transformers>=2.2.0

# RQ1: Semantic Evaluation
bert-score>=0.3.13

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Logging and Output
rich>=13.0.0
tabulate>=0.9.0

